Rename reporter header/cpp
