+++ meson.build
@@ -8,7 +8,6 @@ include_dirs = include_directories('.', 'include')
 headers = files('include/snitch/snitch.hpp',
                 'include/snitch/snitch_append.hpp',
                 'include/snitch/snitch_capture.hpp',
-                'include/snitch/snitch_catch2_xml.hpp',
                 'include/snitch/snitch_cli.hpp',
                 'include/snitch/snitch_concepts.hpp',
                 'include/snitch/snitch_console.hpp',
@@ -28,10 +27,11 @@ headers = files('include/snitch/snitch.hpp',
                 'include/snitch/snitch_macros_warnings.hpp',
                 'include/snitch/snitch_matcher.hpp',
                 'include/snitch/snitch_registry.hpp',
+                'include/snitch/snitch_reporter_catch2_xml.hpp',
+                'include/snitch/snitch_reporter_teamcity.hpp',
                 'include/snitch/snitch_section.hpp',
                 'include/snitch/snitch_string.hpp',
                 'include/snitch/snitch_string_utility.hpp',
-                'include/snitch/snitch_teamcity.hpp',
                 'include/snitch/snitch_test_data.hpp',
                 'include/snitch/snitch_type_name.hpp',
                 'include/snitch/snitch_vector.hpp')
