+++ src/snitch.cpp
@@ -1,13 +1,13 @@
 #include "snitch_append.cpp"
 #include "snitch_capture.cpp"
-#include "snitch_catch2_xml.cpp"
 #include "snitch_cli.cpp"
 #include "snitch_console.cpp"
 #include "snitch_error_handling.cpp"
 #include "snitch_main.cpp"
 #include "snitch_matcher.cpp"
 #include "snitch_registry.cpp"
+#include "snitch_reporter_catch2_xml.cpp"
+#include "snitch_reporter_teamcity.cpp"
 #include "snitch_section.cpp"
 #include "snitch_string_utility.cpp"
-#include "snitch_teamcity.cpp"
 #include "snitch_test_data.cpp"
